requirements.txt
aiogram>=3,<4
SQLAlchemy>=2,<3
pydantic>=2,<3
python-dotenv>=1,<2
python-dateutil>=2,<3
rapidfuzz>=3,<4
pytest>=7,<9
alembic>=1,<2
aiohttp>=3,<4
requests>=2,<3
Faker>=24,<26
